!Hell PRO Tanki Launcher 1.0.5.56 Beta 3

+++ !Hell PRO Tanki Launcher/Program.cs
@@ -29,18 +29,22 @@ namespace _Hell_PRO_Tanki_Launcher
                 try
                 {
                     // Если в папке лежит уже скачанное обновление, то применяем его,
-                    if (File.Exists("launcher.update"))
+                    // а перед этим, на всякий случай, сверяем версии файлов
+                    if (File.Exists("launcher.update") && new Version(FileVersionInfo.GetVersionInfo("launcher.update").FileVersion) > new Version(Application.ProductVersion))
                     {
-                        // а перед этим, на всякий случай, сверяем версии файлов
-                        if (new Version(FileVersionInfo.GetVersionInfo("launcher.update").FileVersion) > new Version(Application.ProductVersion))
-                        {
-                            Process.Start("updater.exe", "launcher.update \"" + Process.GetCurrentProcess().ProcessName + "\"");
-                            Process.GetCurrentProcess().Kill();
-                        }
+                        Process.Start("updater.exe", "launcher.update \"" + Process.GetCurrentProcess().ProcessName + "\"");
+                        Process.GetCurrentProcess().Kill();
+                    }
+                    else
+                    {
+                        if (File.Exists("launcher.update")) { File.Delete("launcher.update"); }
                     }
                 }
                 catch (Exception ex)
                 {
+                    // Если файл поврежден, то удаляем его
+                    if (File.Exists("launcher.update")) { File.Delete("launcher.update"); }
+
                     debug.Save("static void Main()", "if (File.Exists(\"launcher.update\"))", ex.Message);
                 }
